requirements.txt
aws-cdk-lib==2.13.0
constructs>=10.0.0,<11.0.0
boto3==1.21.3
botocore==1.24.3
jmespath==0.10.0
python-dateutil==2.8.2
s3transfer==0.5.1
six==1.16.0
urllib3==1.26.8
python-dotenv==0.19.2
gspread==5.1.1
